% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMAP.R
\name{DEMAP}
\alias{DEMAP}
\title{DEMAP}
\usage{
DEMAP(LogPostLike, control_pars = AlgoParsDEMAP(), ...)
}
\arguments{
\item{LogPostLike}{function whose first arguement is an n_pars-dimensional model parameter vector and returns (scalar) sum of log prior density and log likelihood for the parameter vector.}

\item{control_pars}{control parameters for DE algo. see \code{\link{AlgoParsDEMAP}} function documentation for more details.}

\item{...}{additional arguments to pass LogPostLike}
}
\value{
list contain posterior samples from DEMCMC in a n_samples_per_chain $x$ n_chains $x$ n_pars array and the log likelihood of each sample in a n_samples_per_chain x n_chains array.
}
\description{
DE optimization for Maximum a posteriori estimation; his function tries to find the posterior mode.
}
\examples{
#simulate from model
dataExample=matrix(stats::rnorm(100,c(-1,1),c(1,1)),nrow=50,ncol=2,byrow = TRUE)
#list parameter names
par_names_example=c("mu_1","mu_2")

#log posterior likelihood function = log likelihood + log prior | returns a scalar
LogPostLikeExample=function(x,data,par_names){
 out=0

 names(x)<-par_names

 # log prior
 out=out+sum(dnorm(x["mu_1"],0,sd=1,log=TRUE))
 out=out+sum(dnorm(x["mu_2"],0,sd=1,log=TRUE))

 # log likelihoods
 out=out+sum(dnorm(data[,1],x["mu_1"],sd=1,log=TRUE))
 out=out+sum(dnorm(data[,2],x["mu_2"],sd=1,log=TRUE))

 return(out)
}

# Get map estimattes
DEMAP(LogPostLike=LogPostLikeExample,
      control_pars=AlgoParsDEMCMC(n_pars=length(par_names_example),
                                  n_iter=1000,
                                  n_chains=12),
                                  data=dataExample,
                                  par_names = par_names_example)
}
