% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMCMC.R
\name{DEMCMC}
\alias{DEMCMC}
\title{DEMCMC}
\usage{
DEMCMC(LogPostLike, control_pars = AlgoParsDEMCMC(), ...)
}
\arguments{
\item{name}{LogPostLike}
}
\value{
posteriors samples from DEMCMC
}
\description{
DEMCMC
}
\examples{
#simulate from model
dataExample=matrix(rnorm(100,c(-1,1),c(1,1)),nrow=50,ncol=2,byrow = T)
#list parameter names
par_names_example=c("mu_1","mu_2")

#log posterior likelihood function = log likelihood + log prior | returns a scalar
LogPostLikeExample=function(x,data,par_names){
 out=0

 names(x)<-par_names

 # log prior
 out=out+sum(dnorm(x["mu_1"],0,sd=1,log=T))
 out=out+sum(dnorm(x["mu_2"],0,sd=1,log=T))

 # log likelihoods
 out=out+sum(dnorm(data[,1],x["mu_1"],sd=1,log=T))
 out=out+sum(dnorm(data[,2],x["mu_2"],sd=1,log=T))

 return(out)
}

# Sample from posterior
DEMCMC(LogPostLike=LogPostLikeExample,
      control_pars=AlgoParsDEMCMC(n_pars=length(par_names_example),
                                  n_iter=1000, n_chains=12,
                                  init_sd=.01,init_center=0,
                                  n_cores_use=1,step_size=NULL,
                                  jitter_size=1e-6,parallelType = "none"),data=dataExample,
      par_names = par_names_example)
}
